Working on the probes.
Trying to do the right thing! The CTASensor has information about the sensor but nothing else. A probe on the other hand has a sensor, a calibration, a directional calibration, a hardware, support and cable. So I created the following structs to handle these things:

 - HWSupport
 - HWCable
 - HWBridge
 - Probe1d
 - Probe3d

(Probe2d will come later...)