Added constructors for `Resistor` and `Thermistor` with only keyword arguments
Might be usefull in some cases. And the keyword arguments don't use unicode input. So it uses R0, T0 and a (instead of alpha).

I just tested and (obviously) the code doesn't work with Unitful objects. It might be an interesting feature to have (and a nice Julia programming exercise...). For now I don't think this is useful.