Started to work on CTASensor (ctasensor.jl)
I think I found the correct structure for a flexible framework.

The real issue is that up to now, CTASensor (and CCASensor) really didn't
do anything! It just stored The operating resistance (and current for CCA).

The idea now, is that CTASensor (and CCASensor) will handle most of the
low level stuff, it will actually implement a proxy for a sensor!

It now has besides the operating conditions (Rw for CTA and I for CCA),
the calibration and correction procedures. The calibration will
simply store the calibration curve and calibration conditions.

The correction procedure will correct the readings due to changes in
temperature, pressure and even fluid. It migh even take into account
the heat transfer characteristics of the heating element - important when
dealing with thermistors that hava an insulating layer and thick terminals
where heat is dissipated.

Now I need to get to work on this...