Going back in the interface
Now, the correction and curve fit are part of the CTASensor and CCASensor.
In the end this is simpler...

I think I will change the `correct` method to return the
kinematic viscosity. It is simpler and probably more efficient.

I think I will also return the correction factor instead of the corrected
output. It might be more useful when using the code efficiently in long
time series.